refactor(fsd): A2 F1 — src/ 直下を規約形へ（tests 移設・CSS を shared/ui/theme へ）
A2（#409）の第1波。規約 01 §1-1 は src/ 直下を「5層＋main.tsx のみ」と定めるが、
余分な住人が5件あった。うち3件を解消する。

- src/test/ → frontend/tests/（規約 §2 のテスト基盤置き場）。@tests alias を
  vite/vitest/tsconfig の3箇所に配線し、8ファイルの @/test import を repoint。
  tsconfig の include に tests を追加（型検査から外れないように）
- src/index.css → src/shared/ui/theme/index.css、src/styles/design.css →
  同 theme/design.css。規約 §2 の「main.tsx が shared/ui/theme/index.css を
  import する唯一の場所」の形へ。@layer の宣言順は不変

残る2件はいずれも裁定済みで本波の射程外:
- src/components/keyboard/ … fleet-tooling#89 で凍結中（#385 で追跡）
- src/vite-env.d.ts … hub 裁定 Q2（Vite の標準生成物・条文は手書きソースの規定と読む）

🔴 knip が tailwindcss を「未使用」と誤検出したので、原因を特定して直した:
knip は **src/index.css という慣例パスをハードコードで見ており**、CSS が動くと
その中の `@import "tailwindcss"` を見失う。alias か相対 import かは無関係だった
（両方で再現・切り分け済み）。**ignore で黙らせるのではなく entry/project を
移動先へ向け直した**。陽性対照つき: tailwind の @import を外すと再び検出する＝
検出器は生きたまま。この区別は knip.json のコメントにも残した
（schema.gen.ts の ignore は「握り潰し」であり別物・A2 F2 で削除予定）。

🔴 挙動不変の機械的証明（純移設なので成果物が一致するはず、を実測）:
  CSS  md5 6f4b52865ba3bd2e93d78d89ca951b5e / 477,675 bytes  before == after ✅
  JS   md5 0b13201641606ee538b1170f60548804                   before == after ✅
CSS の md5 は **本番稼働中のバンドル**（current.md に記録された 2026-07-24 デプロイ）
とも一致する。

実測: npm run check 全緑（codegen:check / type-check / eslint / vitest 92 / knip）＋
npm run build 成功。

Refs #409, #385
