+++ frontend/src/main.tsx
@@ -4,7 +4,7 @@ import '@fontsource/noto-sans-jp/600.css'
 import '@fontsource/noto-sans-jp/700.css'
 import '@fontsource/inter/400.css'
 import '@fontsource/inter/600.css'
-import './index.css'
+import '@/shared/ui/theme/index.css'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from '@/app/App'
