+++ frontend/src/shared/ui/theme/design.css
