+++ frontend/vitest.config.ts
@@ -8,12 +8,17 @@ const __dirname = dirname(fileURLToPath(import.meta.url))
 export default defineConfig({
   plugins: [react()],
   resolve: {
-    alias: { '@': resolve(__dirname, 'src') },
+    alias: {
+      '@': resolve(__dirname, 'src'),
+      // Test infrastructure lives outside src/ (規約 01 §2): src/ holds only the
+      // five FSD layers plus main.tsx.
+      '@tests': resolve(__dirname, 'tests'),
+    },
   },
   test: {
     environment: 'jsdom',
     globals: true,
-    setupFiles: ['./src/test/setup.ts'],
+    setupFiles: ['./tests/setup.ts'],
     css: false,
     include: ['src/**/*.{test,spec}.{ts,tsx}'],
   },
